refactoring location coordinates

+++ data/src/main/java/androidbox/me/entities/request/ForecastRequestEntity.kt
@@ -1,7 +1,7 @@
 package androidbox.me.entities.request
 
-data class ForecastRequestEntity(val latitude: Float,
-                                 val longitude: Float,
+data class ForecastRequestEntity(val latitude: Double,
+                                 val longitude: Double,
                                  val days: Int)
